fix broken tzfix

+++ svc/init.d/01-tzfix.sh
@@ -1,6 +1,11 @@
-#!/system/secure/su /bin/sh
+#!/mnt/secure/su /bin/sh
 # This is necessary, otherwise processes with /etc/localtime open will get killed on USB storage.
-mount -o remount,rw /
-rm -f /etc/localtime
-cp -f /mnt/ext1/system/timezone /etc/localtime
-chmod 644 /etc/localtime
+# Don't bother with write-free shadowing, localtime is too trivial for that.
+f=`readlink /etc/localtime`
+if [ "$f" != "" ]; then
+        mount -o remount,rw /
+        rm -f /etc/localtime
+        cp -f "$f" /etc/localtime
+        chmod 644 /etc/localtime
+        mount -o remount,ro /
+fi
